Now, we only have a single example project
