+++ LeanTest.sln
@@ -15,8 +15,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{62FABEA8
 		Build.cmd = Build.cmd
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "State.Examples.MsTest", "Source\State.Examples.MsTest\State.Examples.MsTest.csproj", "{A7470C2D-CB2B-4D63-BEE2-835381A29F00}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSTest", "Source\MSTest\MSTest.csproj", "{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{67FE6255-20F5-443D-979A-8839C6666C9B}"
@@ -36,7 +34,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "LiterateCS", "LiterateCS", 
 		TOC.yml = TOC.yml
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Examples.L0Tests", "Source\Core.Examples.L0Tests\Core.Examples.L0Tests.csproj", "{5641D250-DBDD-41C9-8A08-0E90AE5F3B7F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples.L0Tests", "Source\Examples.L0Tests\Examples.L0Tests.csproj", "{5641D250-DBDD-41C9-8A08-0E90AE5F3B7F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,10 +54,6 @@ Global
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Release|Any CPU.ActiveCfg = Release|Any CPU
