+++ Source/Examples.L0Tests/Application/CompositionRoot.cs
