+++ Source/Examples.L0Tests/Domain/MyOtherData.cs
