+++ Source/Examples.L0Tests/Mocks/MockMyExternalService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Core.Examples.L0Tests.Application;
 using Core.Examples.L0Tests.Domain;
@@ -7,6 +8,7 @@ using LeanTest.Mock;
 
 namespace Core.Examples.L0Tests.Mocks
 {
+	[SuppressMessage("ReSharper", "ClassNeverInstantiated.Global", Justification = "Instantiated by the IoC container")]
 	internal class MockMyExternalService : IMockForData<MyData>, IMockForData<MyOtherData>, IMyExternalService
 	{
 		private List<MyData> MyData { get; } = new List<MyData>();
