+++ Source/Examples.L0Tests/TestSetup/AssemblyInitializer.cs
