+++ Source/Examples.L0Tests/TestSetup/IoC/L0CompositionRootForTest.cs
@@ -1,32 +1,22 @@
-using System;
-using System.Collections.Generic;
 using Core.Examples.L0Tests.Application;
 using Core.Examples.L0Tests.Domain;
 using Core.Examples.L0Tests.Mocks;
+using Core.Examples.L0Tests.StateHandlers;
 using LeanTest.Core.ExecutionHandling;
 using LeanTest.Mock;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Core.Examples.L0Tests.TestSetup.IoC
 {
-	public class IocContainer : IIocContainer
-	{
-		private readonly IServiceProvider _serviceProvider;
-		public IocContainer(IServiceCollection serviceCollection) => _serviceProvider = serviceCollection.BuildServiceProvider();
- 
-		public T Resolve<T>() where T : class => _serviceProvider.GetRequiredService<T>();
-		public T TryResolve<T>() where T : class => _serviceProvider.GetService<T>();
-		public IEnumerable<T> TryResolveAll<T>() where T : class => _serviceProvider.GetServices<T>();
-	}
 	public static class L0CompositionRootForTest
 	{
 		public static IServiceCollection Initialize(IServiceCollection serviceCollection)
 		{
-			// Mocks (not mock-for-data):
-			// TODO
- 
 			// Mock-for-data:
 			serviceCollection.RegisterMockForData<IMyExternalService, MockMyExternalService, MyData, MyOtherData>();
+ 
+			// State handlers:
+			serviceCollection.RegisterStateHandler<MyStateHandler, MyData, MyOtherData>();
  
 			return serviceCollection;
 		}
@@ -39,6 +29,14 @@ namespace Core.Examples.L0Tests.TestSetup.IoC
 			container.AddSingleton<TInterface>(x => x.GetRequiredService<TImplementation>());
 			container.AddSingleton<IMockForData<TData1>>(x => x.GetRequiredService<TImplementation>());
 			container.AddSingleton<IMockForData<TData2>>(x => x.GetRequiredService<TImplementation>());
+		}
+
+		private static void RegisterStateHandler<TImplementation, TData1, TData2>(this IServiceCollection container)
+			where TImplementation: class, IStateHandler<TData1>, IStateHandler<TData2>
+		{
+			container.AddSingleton<TImplementation>();
+			container.AddSingleton<IStateHandler<TData1>>(x => x.GetRequiredService<TImplementation>());
+			container.AddSingleton<IStateHandler<TData2>>(x => x.GetRequiredService<TImplementation>());
 		}
 	}
 }
