--- Source/State.Examples.MsTest/Application/IMyDataLayer.cs
@@ -1,7 +1,0 @@
-namespace State.Examples.MsTest.Application
-{
-	public interface IMyDataLayer
-	{
-		int ReadAgeRecord(string key);
-	}
-}
