--- Source/State.Examples.MsTest/Domain/MyData.cs
@@ -1,8 +1,0 @@
-﻿namespace State.Examples.MsTest.Domain
-{
-	public class MyData
-	{
-		public string Key { get; set; }
-		public int Age { get; set; }
-	}
-}
