--- Source/State.Examples.MsTest/Domain/MyOtherData.cs
@@ -1,8 +1,0 @@
-﻿namespace State.Examples.MsTest.Domain
-{
-	public class MyOtherData
-	{
-		public string OtherKey { get; set; }
-		public int OtherAge { get; set; }
-	}
-}
